collision detection

+++ src/control/include/custom_joint_publisher.h
@@ -33,18 +33,18 @@ float CLOSE_GRIP = 2.5;
 float UP_HEIGHT = 0.55;
 float DOWN_HEIGHT = 0.71;
 int NUMBER_OF_CLASSES = 11;
-Eigen::Vector3f FINAL_POSITIONS[] = {
-    Eigen::Vector3f(0.65, 0.55, 1),
-    Eigen::Vector3f(0.65, 0.4, 1),
-    Eigen::Vector3f(0.8, 0.7, 1),
-    Eigen::Vector3f(0.65, 0.25, 1),
-    Eigen::Vector3f(0.8, 0.55, 1),
-    Eigen::Vector3f(0.8, 0.25, 1),
-    Eigen::Vector3f(0.8, 0.4, 1),
-    Eigen::Vector3f(0.95, 0.7, 1),
-    Eigen::Vector3f(0.95, 0.55, 1),
-    Eigen::Vector3f(0.95, 0.25, 1),
-    Eigen::Vector3f(0.95, 0.4, 1),
+Eigen::Vector3d FINAL_POSITIONS[] = {
+    Eigen::Vector3d(0.65, 0.55, 1),
+    Eigen::Vector3d(0.65, 0.4, 1),
+    Eigen::Vector3d(0.8, 0.7, 1),
+    Eigen::Vector3d(0.65, 0.25, 1),
+    Eigen::Vector3d(0.8, 0.55, 1),
+    Eigen::Vector3d(0.8, 0.25, 1),
+    Eigen::Vector3d(0.8, 0.4, 1),
+    Eigen::Vector3d(0.95, 0.7, 1),
+    Eigen::Vector3d(0.95, 0.55, 1),
+    Eigen::Vector3d(0.95, 0.25, 1),
+    Eigen::Vector3d(0.95, 0.4, 1),
 };
 
 // Publishers
